Removed old 'reports' files (views and policy), and moved the reports out of the 'Organizations' namespace
