--- app/policies/organizations/report_policy.rb
@@ -1,9 +1,0 @@
-module Organizations
-  class ReportPolicy < ApplicationPolicy
-    %i[ index detailed ].each do |action|
-      define_method("#{action}?") do
-        true
-      end
-    end
-  end
-end
