+++ app/views/layouts/_navbar.html.erb
@@ -9,7 +9,7 @@
         <% if current_user.organization_admin? %>
           <%= link_to t("common.workspace"), workspace_projects_path, class: class_names("h-full py-6 lg:py-8 transition duration-300 ease-in-out", is_page_active?([workspace_team_members_path, invite_users_workspace_team_members_path, workspace_projects_path, new_workspace_project_path, edit_workspace_project_path(@project&.id || 0), workspace_project_path(@project&.id || 0), workspace_client_path(@client&.id || 0)]) ? "text-primary font-semibold border-b-2 border-primary": "hover:text-primary text-gray-600") %>
         <% end %>
-        <%= link_to t("common.reports"), organizations_reports_path, class: class_names("h-full py-6 lg:py-8 transition duration-300 ease-in-out", is_page_active?([organizations_reports_path, organizations_detailed_reports_path]) ? "text-primary font-semibold border-b-2 border-primary": "hover:text-primary text-gray-600") %>
+        <%= link_to t("common.reports"), reports_path, class: class_names("h-full py-6 lg:py-8 transition duration-300 ease-in-out", is_page_active?([reports_path, detailed_reports_path]) ? "text-primary font-semibold border-b-2 border-primary": "hover:text-primary text-gray-600") %>
       </div>
       <div class="flex items-center divide-x divide-gray-100 space-x-4">
         <%= render partial: "shared/language_picker" %>
@@ -34,7 +34,7 @@
                   <%= render PhlexUI::Label.new { t("common.workspace")} %>
                 </div>
               <% end %>
-              <%= render PhlexUI::Link.new(href: organizations_reports_path, variant: :ghost, class: "w-full hover:bg-gray-100 flex justify-start pl-2 pr-4") do %>
+              <%= render PhlexUI::Link.new(href: reports_path, variant: :ghost, class: "w-full hover:bg-gray-100 flex justify-start pl-2 pr-4") do %>
                 <div class="w-full flex justify-start items-center gap-x-1">
                   <%= render PhlexUI::Label.new { t("common.reports") } %>
                 </div>
