--- app/views/organizations/reports/details/_extended.html.erb
@@ -1,100 +1,0 @@
-<div class="w-full p-4 print:p-0 flex flex-col gap-y-8 lg:gap-y-16 print:gap-y-4">
-  <div class="flex flex-col">
-    <div class="border-b border-gray-100 pb-4 mb-4 print:hidden">
-      <span class="text-sm print:text-xs font-medium text-gray-500 uppercase"><%= t("common.summary")%></span>
-    </div>
-    <div class="flex-1 flex gap-x-16 mb-4">
-      <div class="flex flex-row items-start gap-x-2">
-        <i class="uc-icon text-2xl text-gray-400 print:hidden">&#xea95;</i>
-        <div>
-          <h2 class="text-base print:text-sm font-regular"><%= t("common.total_billable_hours")%></h2>
-          <span class="font-semibold text-2xl print:text-lg text-primary print:text-gray-900">
-            <%= minutes_to_float detailed_report_data.total_billable_minutes %>
-          </span>
-        </div>
-      </div>
-      <div class="flex flex-row items-start gap-x-2">
-        <i class="uc-icon text-2xl text-gray-400 print:hidden">&#xeb5b;</i>
-        <div>
-          <h2 class="text-base print:text-sm font-regular"><%= t("common.total_hours")%></h2>
-          <span class="font-semibold text-2xl print:text-lg text-primary print:text-gray-900">
-            <%= minutes_to_float detailed_report_data.total_minutes %>
-          </span>
-        </div>
-      </div>
-    </div>
-  </div>
-
-  <div class="flex flex-col">
-    <div class="border-b border-gray-100 pb-4 mb-4 print:mb-0 print:border-none">
-      <span class="text-sm print:text-xs font-medium text-gray-500 uppercase"><%= t("common.time_registrations")%></span>
-    </div>
-    <table class="w-full">
-      <thead>
-      <tr class="bg-gray-50 border-y-2 border-gray-100 text-gray-500 print-color-adjust uppercase font-medium text-base print:text-xs">
-        <td class="pl-5 py-4 print:pl-1 print:py-1"><%= t("common.client")%> (<%= detailed_report_data.clients.count %>)</td>
-        <td><%= t("common.project")%> (<%= detailed_report_data.projects.count %>)</td>
-        <td><%= t("common.task")%> (<%= detailed_report_data.tasks.count %>)</td>
-        <td><%= t("common.rate")%> (<%= detailed_report_data.tasks.count %>)</td>
-        <td><%= t("common.person")%> (<%= detailed_report_data.users.count %>)</td>
-        <td class="w-[6rem] text-end"><%= t("common.hours")%></td>
-        <td class="w-6"></td>
-      </tr>
-      </thead>
-      <tbody class="print:text-sm">
-        <% structured_report_data.each do |key, value| %>
-          <tr class="bg-gray-50 border-b-1 border-gray-100 text-gray-600 print-color-adjust">
-            <td class="pl-5 py-4 print:pl-1 print:py-1 font-semibold col-span-4"><%= key %></td>
-            <td></td>
-            <td></td>
-            <td></td>
-            <td></td>
-            <td class="text-end font-medium"><%= minutes_to_float value.sum { |v| v.minutes } %></td>
-            <td></td>
-          </tr>
-          <% value.each do |child| %>
-            <tr class="text-base print:text-xs border-b last:border-b-0 border-gray-100">
-              <td class="pl-5 print:pl-1 py-4 print:py-1"><%= child.project.client.name %></td>
-              <td>
-                <%= child.project.name %>
-              </td>
-              <td>
-                <div class="flex flex-col gap-3">
-                  <div class="flex flex-col">
-                    <span><%= child.task.name %></span>
-                    <%= simple_format child.notes, wrapper_tag: "span", class: "text-gray-500 text-sm print:text-xs" %>
-                  </div>
-                </div>
-              </td>
-              <td>
-                <div class="flex flex-col">
-                  <%= render PhlexUI::Badge.new(size: :sm, variant: child.assigned_task.rate.zero? ? :gray : :purple, class: "text-nowrap w-fit gap-x-1 justify-center") do %>
-                    <span><%= child.assigned_task.rate.zero? ? t("common.project_rate") : t("common.custom_rate") %></span>
-                  <% end %>
-                  <span class="text-s block">NOK <%= child.assigned_task.rate.zero? ? child.project.rate_nok : child.assigned_task.rate_nok %> </span>
-                </div>
-              </td>
-              <td><%= child.user.name %></td>
-              <td class="text-end"><%= minutes_to_float child.minutes %></td>
-              <td></td>
-            </tr>
-          <% end %>
-        <% end %>
-      </tbody>
-      <tfoot class="border-t border-gray-100 pt-4">
-        <tr class="py-4 print:py-2">
-          <td></td>
-          <td></td>
-          <td></td>
-          <td></td>
-          <td></td>
-          <td class="text-end">
-            <span class="font-medium text-xl print:text-lg"><%= t("common.total")%></span>
-            <span class="font-semibold text-2xl print:text-xl text-primary print:text-gray-900"><%= minutes_to_float detailed_report_data.total_minutes %></span>
-          </td>
-          <td></td>
-        </tr>
-      </tfoot>
-    </table>
-  </div>
-</div>
