--- app/views/organizations/reports/index.html.erb
@@ -1,2 +1,0 @@
-<%= render partial: "organizations/reports/top_nav/index" %>
-<%= render partial: "organizations/reports/content" %>
