--- app/views/reports/_report_data.html.erb
@@ -1,57 +1,0 @@
-<table class="w-full">
-  <thead class="w-fill">
-    <tr class="bg-gray-200 text-left">
-      <% header_columns = {
-          "date" => "Date",
-          "client" => "Client",
-          "project" => "Project",
-          "task" => "Task",
-          "user" => "Person"
-      } %>
-      <!--iterates over every header column-->
-      <% header_columns.each do |group, title| %>
-        <% unless @report.group_by == group %> <!--prevents a column with the current group_by -->
-          <th class="border-b-2 border-gray-300 p-1"><%= title %></th>
-        <% end %>
-      <% end %>
-      <th class="border-b-2 border-gray-300">hours</th>
-    </tr>
-  </thead>
-  <tbody class="w-full">
-    <!--iterates over each group with time_regs -->
-    <% grouped_report.each do | group, time_regs| %>
-
-      <!--creates a row with a group name column fills the rest of the column blank-->
-      <tr class="bg-gray-100">
-        <td class="p-2 text-sm font-semibold  border-b-2 border-gray-300"><%= group %><br></td>
-        <% 4.times do %>
-          <td class="border-b-2 border-gray-300 p-1"></td>
-        <% end %>
-      </tr>
-
-      <!-- iterates over each time_reg in each group -->
-      <% time_regs.each do |time_reg |%>
-        <tr class="border-b-2 border-gray-300">
-          <%  content_columns = {
-                "date" => time_reg[:date],
-                "client" => time_reg[:client],
-                "project" => time_reg[:project],
-                "task" => time_reg[:task],
-                "user" => time_reg[:user]
-          }%>
-          <% content_columns.each do |group, text |%>
-            <% unless @report.group_by == group %> <!--prevents a column with the current group_by -->
-              <td class="pb-1 p-1"><%= text %></td>
-            <% end %>
-          <% end %>
-          <td class="pb-1"><%= convert_time_int(time_reg[:minutes])%></td>
-        </tr>
-      <% end %>
-      <!-- filler-row at bottom to space out the groups -->
-      <tr class="">
-        <td><br></td>
-      </tr>
-    <% end %>
-  </tbody>
-</table>
-
