+++ app/views/reports/_row.html.erb
@@ -2,7 +2,7 @@
   <%= render PhlexUI::Table::Cell.new do %>
     <%# TODO: Refactor this and find a way to persist old unchanged query params %>
     <%- disable_link = @filter.visible_tabs.length <= 1 -%>
-    <%= link_to disable_link ? "#" : organizations_reports_path(filter: @filter.to_h(**result[:group_link])), class: class_names("flex flex-col", disable_link ? "cursor-default" : "underline text-blue-600") do %>
+    <%= link_to disable_link ? "#" : reports_path(filter: @filter.to_h(**result[:group_link])), class: class_names("flex flex-col", disable_link ? "cursor-default" : "underline text-blue-600") do %>
       <div class="flex flex-row items-center gap-x-1">
         <span class="font-medium text-sm"><%= result[:attribute_name] %></span>
       </div>
