+++ app/views/reports/_summary.html.erb
