+++ app/views/reports/details/_filter.html.erb
@@ -1,5 +1,5 @@
 <div class="bg-gray-100 border-b py-4 border-gray-200 px-4 print:border-0">
-  <%= form_with scope: :filter, url: organizations_detailed_reports_path, method: :get, data: { controller: :form }, class: "print:hidden" do |form| %>
+  <%= form_with scope: :filter, url: detailed_reports_path, method: :get, data: { controller: :form }, class: "print:hidden" do |form| %>
     <div class="max-w-7xl mx-auto flex justify-between items-center">
       <div class="flex flex-col gap-y-2">
         <span class="uppercase text-sm text-gray-600 font-semibold"><%= t('common.filter') %></span>
@@ -104,7 +104,7 @@
             <%# team end %>
           </div>
           <div class="flex flex-row items-center gap-x-4">
-            <%= link_to organizations_detailed_reports_path, class: "text-xs text-gray-400 underline text-nowrap" do %>
+            <%= link_to detailed_reports_path, class: "text-xs text-gray-400 underline text-nowrap" do %>
               <span><%= t('common.reset') %></span>
               <i class="uc-icon">&#xeb8c;</i>
             <% end %>
