--- app/views/reports/details/_no_results.html.erb
@@ -1,17 +1,0 @@
-<div data-controller="report-edit" class="flex flex-col p-4 print:p-0">
-  <div class="w-full">
-    <div class="flex flex-col items-center justify-center py-8 lg:py-16 gap-y-4">
-      <%= image_tag "empty.svg" %>
-      <div class="gap-y-2 flex flex-col text-center w-full md:w-3/4 lg:w-1/2 mx-auto">
-        <span class="font-semibold text-lg"><%= t('common.no_time_registrations') %></span>
-        <span class="text-base text-gray-500"><%= t('common.No_registrations_were_found') %></span>
-      </div>
-      <div>
-        <%= render ButtonComponent.new(path: report_path, method: :get, variant: :secondary) do %>
-          <span class="mr-2"><%= t('common.reset_filter') %></span>
-          <i class="uc-icon text-xl">&#xeb8c;</i>
-        <% end %>
-      </div>
-    </div>
-  </div>
-</div>
