+++ app/views/reports/index.html.erb
@@ -1,0 +1,2 @@
+<%= render partial: "reports/top_nav/index" %>
+<%= render partial: "reports/content" %>
