+++ app/views/reports/top_nav/_selected_time_range.html.erb
